requirements.txt
# Web framework
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
aiosqlite==0.20.0

# Settings
pydantic-settings==2.5.0

# Face recognition
insightface==0.7.3
onnxruntime==1.19.0
opencv-python-headless==4.10.0.84
numpy==1.26.4

# Vector database
pinecone-client==5.0.1

# Dashboard
streamlit==1.38.0
requests==2.32.3

# Testing
pytest==8.3.3
httpx==0.27.2
